+++ CHANGELOG.rst
@@ -124,3 +124,4 @@ Changelog
   difficulty and character attribute (strength, perception, intelligence,
   constitution)
 * Habitica cards are now created with a "Trello" tag.
+* Added due date synchronisation support.